refactor(ffx-orm): rename file too!
